splitted CveStore into two classes
so I can reuse some code
